% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed9_to_granges.R
\name{bed9_to_granges}
\alias{bed9_to_granges}
\title{Converts imported bed file to granges object}
\usage{
bed9_to_granges(bed_dataframe, genome = "mm10")
}
\arguments{
\item{bed_dataframe}{Imported bed data.frame() object (from \code{bed9_importer})}

\item{genome}{Which genome is this? Defaults to mm10. Other options are FILL IN!!}
}
\description{
Converts imported bed file to granges object
}
\details{
\code{bed9_to_granges} Converts an imported bed file to a granges
object for annotation.
}
\examples{
bed_granges <- bed9_to_granges(bed9_dataframe, genome='hg19')
bed_granges
}
