% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_bigBed_to_bed.R
\name{convert_bigBed_to_bed}
\alias{convert_bigBed_to_bed}
\title{Wrapper to converted bigBed to bed, using UCSC bedToBigBed tool}
\usage{
convert_bigBed_to_bed(bigBed)
}
\arguments{
\item{bigBed}{Path to the bigBed file}
}
\description{
Wrapper to converted bigBed to bed, using UCSC bedToBigBed tool
}
\details{
You must have the UCSC bedToBigBed tool installed. UCSC provides pre-built binaries here: http://hgdownload.soe.ucsc.edu/admin/exe/. \code{convert_bigBed_to_Bed} will return the path to the converted bed file.
}
\examples{
bed_file_path <- convert_bigBed_to_bed('~/mm10_hela_chromhmm.bigBed')
bed_file_path
}
