% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_annotations.R
\name{load_annotations}
\alias{load_annotations}
\title{Load species and build specific annotations}
\usage{
load_annotations(build = "mm10", type = "gene_body")
}
\arguments{
\item{type}{Gene body is the default. You can also load exon only, promoter
only, intron only.}

\item{bed}{Species and genome build}
}
\description{
Load species and build specific annotations
}
\details{
\code{load_annotations} Loads in species and build specific annotations
encoding (character, integer) of the columns.
}
\examples{
annotations <- load_annotations(build= 'mm10', type='gene_byd' )
annotations
}
